+++ CHANGES.md
@@ -3,6 +3,9 @@
 
 ## 2.0.0 - January 20, 2016
 
+- Added the bufferbuttons plugin
+  [ale-rt]
+
 - Support for redactor II
 
 ## 0.0.1 - September 28, 2015